[#1] - Moved the SanitizerInterface to the Stoic\Input\Sanitizers namespace in order to help keep the root folder uncluttered, and because it just made sense.
